# Reduce heap allocations in `auparse.go`