--- a/auparse/auparse.go
+++ b/auparse/auparse.go
@@ -24,6 +24,7 @@ import (
 	"regexp"
 	"strconv"
 	"strings"
+	"sync"
 	"syscall"
 	"time"
 
@@ -54,7 +55,7 @@ type AuditMessage struct {
 	Sequence   uint32           // Sequence parsed from payload.
 	RawData    string           // Raw message as a string.
 
-	fields map[string]*field
+	fields *fieldMap
 	offset int               // offset is the index into RawData where the header ends and message begins.
 	data   map[string]string // The key value pairs parsed from the message.
 	tags   []string          // The keys associated with the event (e.g. the values set in rules with -F key=exec).
@@ -66,10 +67,45 @@ type field struct {
 	value string // Parsed and enriched value.
 }
 
-func newField(orig string) *field { return &field{orig: orig, value: orig} }
-func (f *field) Orig() string     { return f.orig }
-func (f *field) Value() string    { return f.value }
-func (f *field) Set(value string) { f.value = value }
+func newField(orig string) *field   { return &field{orig: orig, value: orig} }
+func newFieldVal(orig string) field { return field{orig: orig, value: orig} }
+func (f *field) Orig() string       { return f.orig }
+func (f *field) Value() string      { return f.value }
+func (f *field) Set(value string)   { f.value = value }
+
+type fieldMap map[string]field
+
+func (fm *fieldMap) add(key string, f field) {
+	m := *fm
+	m[key] = f
+}
+
+func (fm *fieldMap) setFieldValue(key, value string) {
+	m := *fm
+	if f, ok := m[key]; ok {
+		m[key] = field{f.orig, value}
+	}
+	return
+}
+
+func (fm *fieldMap) find(key string) (field, error) {
+	m := *fm
+	if f, ok := m[key]; ok {
+		return f, nil
+	}
+	return field{}, fmt.Errorf("%s key not found", key)
+}
+
+func (fm *fieldMap) delete(key string) {
+	delete(*fm, key)
+}
+
+// A pool of intermediate maps to relief pressure from the GC
+var fieldPool = sync.Pool{
+	New: func() interface{} {
+		return fieldMap{}
+	},
+}
 
 // Data returns the key-value pairs that are contained in the audit message.
 // This information is parsed from the raw message text the first time this
@@ -92,18 +128,16 @@ func (m *AuditMessage) Data() (map[string]string, error) {
 		return nil, m.error
 	}
 
-	m.fields = map[string]*field{}
-	defer func() { m.fields = nil }()
-	extractKeyValuePairs(message, m.fields)
+	*m.fields = extractKeyValuePairsFm(message)
 
-	if err = enrichData(m); err != nil {
+	if err = m.enrichData(); err != nil {
 		m.error = err
 		return nil, m.error
 	}
-
-	m.data = make(map[string]string, len(m.fields))
-	for k, f := range m.fields {
-		m.data[k] = f.Value()
+	d := *m.fields
+	m.data = make(map[string]string, len(d))
+	for k, f := range d {
+		m.data[k] = f.value
 	}
 
 	return m.data, m.error
@@ -189,6 +223,7 @@ func Parse(typ AuditMessageType, message string) (*AuditMessage, error) {
 		Sequence:   seq,
 		offset:     indexOfMessage(message[end:]),
 		RawData:    message,
+		fields:     &fieldMap{},
 	}
 	return msg, nil
 }
@@ -255,6 +290,13 @@ var (
 	// value pairs.
 	kvRegex = regexp.MustCompile(`([a-z0-9_-]+)=((?:[^"'\s]+)|'(?:\\'|[^'])*'|"(?:\\"|[^"])*")`)
 
+	kvregexPool = sync.Pool{
+		New: func() interface{} {
+			r := kvRegex
+			return r
+		},
+	}
+
 	// avcMessageRegex matches the beginning of SELinux AVC messages to parse
 	// the seresult and seperms parameters.
 	// Example: "avc:  denied  { read } for  "
@@ -290,6 +332,60 @@ func normalizeAuditMessage(typ AuditMessageType, msg string) (string, error) {
 	return msg, nil
 }
 
+func extractKeyValuePairsFm(msg string) fieldMap {
+	data := make(fieldMap, 0)
+	r := kvregexPool.Get().(*regexp.Regexp)
+	defer kvregexPool.Put(r)
+
+	matches := r.FindAllStringSubmatch(msg, -1)
+	for _, m := range matches {
+		key, orig := m[1], m[2]
+		value := trimQuotesAndSpace(orig)
+
+		// Drop fields with useless values.
+		switch value {
+		case "", "?", "?,", "(null)":
+			continue
+		}
+
+		if key == "msg" {
+			for mk, mv := range extractKeyValuePairsFm(value) {
+				data.add(mk, mv)
+			}
+			continue
+		}
+		data[key] = field{orig: orig, value: value}
+	}
+	return data
+}
+
+func extractKeyValuePairsPool(msg string) fieldMap {
+	data := fieldPool.Get().(fieldMap)
+	matches := kvRegex.FindAllStringSubmatch(msg, -1)
+	for _, m := range matches {
+		key := m[1]
+		f := newFieldVal(m[2])
+		f.Set(trimQuotesAndSpace(m[2]))
+
+		// Drop fields with useless values.
+		switch f.Value() {
+		case "", "?", "?,", "(null)":
+			continue
+		}
+
+		if key == "msg" {
+			msgData := extractKeyValuePairsPool(f.value)
+			for mk, mv := range msgData {
+				data.add(mk, mv)
+			}
+			fieldPool.Put(msgData)
+			continue
+		}
+		data[key] = f
+	}
+	return data
+}
+
 func extractKeyValuePairs(msg string, data map[string]*field) {
 	matches := kvRegex.FindAllStringSubmatch(msg, -1)
 	for _, m := range matches {
@@ -311,104 +407,82 @@ func extractKeyValuePairs(msg string, data map[string]*field) {
 	}
 }
 
-func extractKeyValuePairsNew(msg string) map[string]field {
-	data := make(map[string]field, 0)
-	matches := kvRegex.FindAllStringSubmatch(msg, -1)
-	for _, m := range matches {
-		key := m[1]
-		f := *newField(m[2])
-		f.Set(trimQuotesAndSpace(m[2]))
-
-		// Drop fields with useless values.
-		switch f.Value() {
-		case "", "?", "?,", "(null)":
-			continue
-		}
-
-		if key == "msg" {
-			data = extractKeyValuePairsNew(f.Value())
-		} else {
-			data[key] = f
-		}
-	}
-	return data
-}
-
 func trimQuotesAndSpace(v string) string { return strings.Trim(v, `'" `) }
 
 // Enrichment after KV parsing
 
 //nolint:errcheck // Continue enriching even if some fields do not exist.
-func enrichData(msg *AuditMessage) error {
-	normalizeUnsetID("auid", msg.fields)
-	normalizeUnsetID("old-auid", msg.fields)
-	normalizeUnsetID("ses", msg.fields)
+func (msg *AuditMessage) enrichData() error {
+	msg.fields.normalizeUnsetID("auid")
+	msg.fields.normalizeUnsetID("old-auid")
+	msg.fields.normalizeUnsetID("ses")
 
 	// Many message types can have subj field so check them all.
-	parseSELinuxContext("subj", msg.fields)
+	_ = msg.fields.parseSELinuxContext("subj")
 
 	// Normalize success/res to result.
-	result(msg.fields)
+	_ = msg.fields.result()
 
 	// Convert exit codes to named POSIX exit codes.
-	exit(msg.fields)
+	_ = msg.fields.exit()
 
 	// Normalize keys that are of the form key="key=user_command".
-	auditRuleKey(msg)
+	auditRuleKeyNew(msg)
 
-	hexDecode("cwd", msg.fields)
+	_ = msg.fields.hexDecode("cwd")
 
 	switch msg.RecordType {
 	case AUDIT_SECCOMP:
-		if err := setSignalName(msg.fields); err != nil {
+		if err := msg.fields.setSignalName(); err != nil {
 			return err
 		}
 		fallthrough
 	case AUDIT_SYSCALL:
-		if err := arch(msg.fields); err != nil {
+		if err := msg.fields.arch(); err != nil {
 			return err
 		}
-		if err := setSyscallName(msg.fields); err != nil {
+		if err := msg.fields.setSyscallName(); err != nil {
 			return err
 		}
-		if err := hexDecode("exe", msg.fields); err != nil {
+		if err := msg.fields.hexDecode("exe"); err != nil {
 			return err
 		}
 	case AUDIT_SOCKADDR:
-		if err := saddr(msg.fields); err != nil {
+		if err := msg.fields.saddr(); err != nil {
 			return err
 		}
 	case AUDIT_PROCTITLE:
-		if err := hexDecode("proctitle", msg.fields); err != nil {
+		if err := msg.fields.hexDecode("proctitle"); err != nil {
 			return err
 		}
 	case AUDIT_USER_CMD:
-		if err := hexDecode("cmd", msg.fields); err != nil {
+		if err := msg.fields.hexDecode("cmd"); err != nil {
 			return err
 		}
 	case AUDIT_TTY, AUDIT_USER_TTY:
-		if err := hexDecode("data", msg.fields); err != nil {
+		if err := msg.fields.hexDecode("data"); err != nil {
 			return err
 		}
 	case AUDIT_EXECVE:
-		if err := execveArgs(msg.fields); err != nil {
+		if err := msg.fields.execveArgs(); err != nil {
 			return err
 		}
 	case AUDIT_PATH:
-		parseSELinuxContext("obj", msg.fields)
-		hexDecode("name", msg.fields)
+		_ = msg.fields.parseSELinuxContext("obj")
+		_ = msg.fields.hexDecode("name")
 	case AUDIT_USER_LOGIN:
 		// acct only exists in failed logins.
-		hexDecode("acct", msg.fields)
+		_ = msg.fields.hexDecode("acct")
 	}
 
 	return nil
 }
 
-func arch(data map[string]*field) error {
-	field, found := data["arch"]
-	if !found {
-		return errors.New("arch key not found")
+func (fm *fieldMap) arch() error {
+	const key = "arch"
+	field, err := fm.find(key)
+	if err != nil {
+		return err
 	}
 
 	arch, err := strconv.ParseInt(field.Value(), 16, 64)
@@ -416,102 +490,102 @@ func arch(data map[string]*field) error {
 		return fmt.Errorf("failed to parse arch: %w", err)
 	}
 
-	field.Set(AuditArch(arch).String())
+	fm.setFieldValue(key, AuditArch(arch).String())
 	return nil
 }
 
-func setSyscallName(data map[string]*field) error {
-	field, found := data["syscall"]
-	if !found {
-		return errors.New("syscall key not found")
+func (fm *fieldMap) setSyscallName() error {
+	field, err := fm.find("syscall")
+	if err != nil {
+		return err
 	}
 
-	syscall, err := strconv.Atoi(field.Value())
+	syscall, err := strconv.Atoi(field.value)
 	if err != nil {
 		return fmt.Errorf("failed to parse syscall: %w", err)
 	}
 
-	arch, found := data["arch"]
-	if !found {
+	arch, err := fm.find("arch")
+	if err != nil {
 		return errors.New("arch key not found so syscall cannot be translated to a name")
 	}
 
-	if name, found := AuditSyscalls[arch.Value()][syscall]; found {
-		field.Set(name)
+	if name, found := AuditSyscalls[arch.value][syscall]; found {
+		fm.setFieldValue("syscall", name)
 	}
 	return nil
 }
 
-func setSignalName(data map[string]*field) error {
-	field, found := data["sig"]
-	if !found {
-		return errors.New("sig key not found")
+func (fm *fieldMap) setSignalName() error {
+	field, err := fm.find("sig")
+	if err != nil {
+		return err
 	}
 
-	signalNum, err := strconv.Atoi(field.Value())
+	signalNum, err := strconv.Atoi(field.value)
 	if err != nil {
 		return fmt.Errorf("failed to parse sig: %w", err)
 	}
 
 	if signalName := unix.SignalName(syscall.Signal(signalNum)); signalName != "" {
-		field.Set(signalName)
+		fm.setFieldValue("sig", signalName)
 	}
 	return nil
 }
 
-func saddr(data map[string]*field) error {
-	field, found := data["saddr"]
-	if !found {
-		return errors.New("saddr key not found")
+func (fm *fieldMap) saddr() error {
+	field, err := fm.find("saddr")
+	if err != nil {
+		return err
 	}
 
-	saddrData, err := parseSockaddr(field.Value())
+	saddrData, err := parseSockaddr(field.value)
 	if err != nil {
 		return fmt.Errorf("failed to parse saddr: %w", err)
 	}
 
-	delete(data, "saddr")
+	fm.delete("saddr")
 	for k, v := range saddrData {
-		data[k] = newField(v)
+		fm.add(k, *newField(v))
 	}
 	return nil
 }
 
-func normalizeUnsetID(key string, data map[string]*field) {
-	field, found := data[key]
-	if !found {
+func (fm *fieldMap) normalizeUnsetID(key string) {
+	f, err := fm.find(key)
+	if err != nil {
 		return
 	}
 
-	switch field.Value() {
+	switch f.Value() {
 	case "4294967295", "-1":
-		field.Set("unset")
+		fm.setFieldValue(key, "unset")
 	}
 }
 
-func hexDecode(key string, data map[string]*field) error {
-	field, found := data[key]
-	if !found {
-		return fmt.Errorf("%v key not found", key)
+func (fm *fieldMap) hexDecode(key string) error {
+	field, err := fm.find(key)
+	if err != nil {
+		return err
 	}
 
 	// Use the original value that may or may not contain a leading quote.
-	decodedStrings, err := hexToStrings(field.Orig())
+	decodedStrings, err := hexToStrings(field.orig)
 	if err != nil {
 		// Field is not in hex. Ignore.
 		return nil
 	}
 
 	if len(decodedStrings) > 0 {
-		field.Set(strings.Join(decodedStrings, " "))
+		fm.setFieldValue(key, strings.Join(decodedStrings, " "))
 	}
 	return nil
 }
 
-func execveArgs(data map[string]*field) error {
-	argc, found := data["argc"]
-	if !found {
-		return errors.New("argc key not found")
+func (fm *fieldMap) execveArgs() error {
+	argc, err := fm.find("argc")
+	if err != nil {
+		return err
 	}
 
 	count, err := strconv.ParseUint(argc.Value(), 10, 32)
@@ -522,13 +596,13 @@ func execveArgs(data map[string]*field) error {
 	for i := 0; i < int(count); i++ {
 		key := "a" + strconv.Itoa(i)
 
-		arg, found := data[key]
-		if !found {
+		arg, err := fm.find(key)
+		if err != nil {
 			return fmt.Errorf("failed to find arg %v", key)
 		}
 
 		if ascii, err := hexToString(arg.Orig()); err == nil {
-			arg.Set(ascii)
+			fm.setFieldValue(key, ascii)
 		}
 	}
 
@@ -537,56 +611,56 @@ func execveArgs(data map[string]*field) error {
 
 // parseSELinuxContext parses a SELinux security context of the form
 // 'user:role:domain:level:category'.
-func parseSELinuxContext(key string, data map[string]*field) error {
-	field, found := data[key]
-	if !found {
-		return fmt.Errorf("%v key not found", key)
+func (fm *fieldMap) parseSELinuxContext(key string) error {
+	f, err := fm.find(key)
+	if err != nil {
+		return err
 	}
 
 	keys := []string{"_user", "_role", "_domain", "_level", "_category"}
-	contextParts := strings.SplitN(field.Value(), ":", len(keys))
+	contextParts := strings.SplitN(f.value, ":", len(keys))
 	if len(contextParts) == 0 {
 		return fmt.Errorf("failed to split SELinux context field %v", key)
 	}
-	delete(data, key)
+	fm.delete(key)
 
 	for i, part := range contextParts {
-		data[key+keys[i]] = newField(part)
+		fm.add(key+keys[i], *newField(part))
 	}
 	return nil
 }
 
-func result(data map[string]*field) error {
+func (fm *fieldMap) result() error {
 	// Syscall messages use "success". Other messages use "res".
-	field, found := data["success"]
-	if !found {
-		field, found = data["res"]
-		if !found {
+	field, err := fm.find("success")
+	if err != nil {
+		field, err = fm.find("res")
+		if err != nil {
 			return errors.New("success and res key not found")
 		}
-		delete(data, "res")
+		fm.delete("res")
 	} else {
-		delete(data, "success")
+		fm.delete("success")
 	}
 
-	switch v := strings.ToLower(field.Value()); {
+	switch v := strings.ToLower(field.value); {
 	case v == "yes", v == "1", strings.HasPrefix(v, "suc"):
-		data["result"] = newField("success")
+		fm.add("result", *newField("success"))
 	default:
-		data["result"] = newField("fail")
+		fm.add("result", *newField("fail"))
 	}
 	return nil
 }
 
-func auditRuleKey(msg *AuditMessage) {
-	field, found := msg.fields["key"]
-	if !found {
+func auditRuleKeyNew(msg *AuditMessage) {
+	field, err := msg.fields.find("key")
+	if err != nil {
 		return
 	}
-	delete(msg.fields, "key")
+	msg.fields.delete("key")
 
 	// Handle hex encoded data (e.g. key=28696E7).
-	if decodedData, err := decodeUppercaseHexString(field.Orig()); err == nil {
+	if decodedData, err := decodeUppercaseHexString(field.orig); err == nil {
 		keys := strings.Split(string(decodedData), string([]byte{0x01}))
 		msg.tags = keys
 		return
@@ -602,10 +676,10 @@ func auditRuleKey(msg *AuditMessage) {
 	}
 }
 
-func exit(data map[string]*field) error {
-	field, found := data["exit"]
-	if !found {
-		return errors.New("exit key not found")
+func (fm *fieldMap) exit() error {
+	field, err := fm.find("exit")
+	if err != nil {
+		return err
 	}
 
 	exitCode, err := strconv.Atoi(field.Value())
@@ -622,6 +696,6 @@ func exit(data map[string]*field) error {
 		return nil
 	}
 
-	field.Set(name)
+	fm.setFieldValue("exit", name)
 	return nil
 }