## Reason for this PR
We noticed high GC pressure (recorded by profiling) when running `auditbeat` and an eBPF-powered profiling agent.
When auditbeat is configured to read _all_ BPF syscalls, and a profiling agent based on eBPF executes many, a relevant CPU consumption goes into parsing and sending the increased number of events from auditd.

## Details
With this PR we have a reduction of ~15% of heap allocations in the parse/enrich phases of `AuditMessage`.
The commits contain some additional benchmarks and experiments that lead to the improvement.

The benchmark output
```
# before
BenchmarkExtractKeyValuePairs/bpf_message_current-8         	  113053	      9687 ns/op	    2198 B/op	      33 allocs/op
BenchmarkExtractKeyValuePairs/short_message_current-8       	  292078	      3440 ns/op	    1232 B/op	       9 allocs/op

# after
BenchmarkExtractKeyValuePairs/bpf_message_current-8         	  118813	      8945 ns/op	    2616 B/op	      24 allocs/op
BenchmarkExtractKeyValuePairs/short_message_current-8       	  586594	      3114 ns/op	     913 B/op	       7 allocs/op
```

This is not a definitive solution to CPU usage increase, but it's a step further 😄