@andrewkroh 2022-06-06
Comparing main (48e317eaca4fa4dfde10a84d070a084aed1ddfa6) to this branch using the existing benchmarks:

`go test -bench . -benchtime=10s -benchmem -run NONE ./auparse ./aucoalesce/`

```
benchmark                             old ns/op     new ns/op     delta
BenchmarkParseAuditHeader-10          143           144           +0.56%
BenchmarkParseAuditHeaderRegex-10     540           543           +0.46%
BenchmarkParseLogLine-10              361           358           -0.69%
BenchmarkCoalesceMessages-10          3114          3128          +0.45%

benchmark                             old allocs     new allocs     delta
BenchmarkParseAuditHeader-10          3              3              +0.00%
BenchmarkParseAuditHeaderRegex-10     2              2              +0.00%
BenchmarkParseLogLine-10              4              4              +0.00%
BenchmarkCoalesceMessages-10          13             13             +0.00%

benchmark                             old bytes     new bytes     delta
BenchmarkParseAuditHeader-10          21            21            +0.00%
BenchmarkParseAuditHeaderRegex-10     128           128           +0.00%
BenchmarkParseLogLine-10              374           358           -4.28%
BenchmarkCoalesceMessages-10          2718          2718          +0.00%
```