@inge4pres 2022-06-08
Re-requesting review, I think I addressed all the comments in https://github.com/elastic/go-libaudit/pull/111/commits/3692eb071a55567ecd91a50215f4d542ee4d1f6a, let me know otherwise 🙏🏼