@inge4pres 2022-06-13
> Can you please add an entry to the CHANGELOG.md in the root of the repo.
> 
> > do you usually squash PR commits?
> 
> Yeah, we normally let GH squash it.

CHANGELOG line added in https://github.com/elastic/go-libaudit/pull/111/commits/8c93c65c9230447d747bfd2c7942964f85cb5cb4.
Thanks both for your help 🙏🏼